scripts/requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0

pyodbc>=4.0.0

pandas>=1.3.0
numpy>=1.21.0

# Environment management
python-dotenv>=0.19.0

sentence-transformers>=2.2.0
transformers>=4.20.0
torch>=1.12.0
faiss-cpu>=1.7.0

spacy>=3.4.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0

python-dotenv>=0.19.0
tqdm>=4.64.0

matplotlib>=3.5.0
seaborn>=0.11.0

pytest>=7.0.0
requests>=2.28.0